move search functionality to discovery resource
