+++ src/index.js
@@ -2,6 +2,7 @@ import ContractService from './services/contract-service'
 import IpfsService from './services/ipfs-service'
 import { Attestations } from './resources/attestations'
 import Marketplace from './resources/marketplace'
+import Discovery from './resources/discovery'
 import Users from './resources/users'
 import Messaging from './resources/messaging'
 import Token from './resources/token'
@@ -10,7 +11,8 @@ import store from 'store'
 
 const defaultBridgeServer = 'https://bridge.originprotocol.com'
 const defaultIpfsDomain = 'gateway.originprotocol.com'
-const defaultDiscoveryServer = 'https://discovery.originprotocol.com'
+//const defaultDiscoveryServer = 'https://discovery.originprotocol.com'
+const defaultDiscoveryServer = 'http://localhost'
 const defaultDiscoveryServerPort = '4000'
 const defaultIpfsApiPort = '5002'
 const defaultIpfsGatewayPort = '443'
@@ -53,10 +55,13 @@ class Origin {
     this.marketplace = new Marketplace({
       contractService: this.contractService,
       ipfsService: this.ipfsService,
+      store
+    })
+
+    this.discovery = new Discovery({
       discoveryServer,
       discoveryServerPort,
-      fetch,
-      store
+      fetch
     })
 
     this.users = new Users({
